play buttons instead of the ⁂
Asterism ⁂ used to indicate minor breaks in text, call attention to a
passage, or to separate sub-chapters in a book, but is hardly used
anymore and very linked to literary contents. Trying out something
closer to the more video-related play button. Or should it be a double
play button as in fast forward?